pyproject.toml
[tool.poetry]
name = "qfa-api"
version = "0.0.1"
authors = [
  "Jacopo Margutti <jmargutti@redcross.nl>"
]
description = "QFA API"
readme = "README.md"
repository = "https://github.com/rodekruis/qfa-api"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
azure-ai-ml = "*"
azure-cosmos = "*"
azure-identity = "*"
azure-monitor-opentelemetry-exporter = "*"
azure-search-documents = "*"
clean-text = "*"
python-dotenv = "*"
fastapi = "*"
orjson = "*"
pandas = "*"
torch = "*"
transformers = "*"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
